+++ plugin-nvidia/Cargo.toml
@@ -18,6 +18,7 @@ nvml-wrapper = { version = "0.10.0", features = ["legacy-functions"], optional =
 nvml-wrapper-sys = { version = "0.8.0", optional = true }
 regex = { version = "1.10.6", optional = true }
 serde = { version = "1.0.210", features = ["derive"] }
+tempfile = "3.10"
 
 [lints]
 workspace = true
