+++ plugin-nvidia/src/jetson/mod.rs
@@ -1,0 +1,4 @@
+pub mod metrics;
+pub mod plugin;
+pub mod probe;
+pub mod utils;
